Upstream dropped the optimized frontal 2D algorithm, renamed the header of
the 2D Delaunay kernel, and stopped choosing the Metis algorithm from the
number of partitions. The declaration follows: the choice list of
Mesh.Algorithm loses its entry, GuiActions.cpp and GuiClassify.cpp include
what they need by name (MQuadrangle.h came through the old header), and
the partition dialog no longer sets the algorithm when the count changes.
The three FLTK files upstream touched are gone on this branch; what it
changed in them is the same three things, carried by the declaration.

+++ src/mesh/meshGRegionHxt.cpp
@@ -296,7 +296,7 @@ static HXTStatus Hxt2Gmsh(std::vector<GRegion *> &regions, HXTMesh *m,
 #if 1
   std::vector<std::size_t> numtet(regions.size(), 0);
   for(std::size_t i = 0; i < m->tetrahedra.num; i++) {
-    uint16_t c = m->tetrahedra.color[i];
+    uint32_t c = m->tetrahedra.color[i];
     if(c >= numtet.size()) continue;
     numtet[c]++;
   }
@@ -305,7 +305,7 @@ static HXTStatus Hxt2Gmsh(std::vector<GRegion *> &regions, HXTMesh *m,
     regions[c]->mesh_vertices.reserve(numtet[c] / 6);
   }
   for(std::size_t i = 0; i < m->tetrahedra.num; i++) {
-    uint16_t c = m->tetrahedra.color[i];
+    uint32_t c = m->tetrahedra.color[i];
     if(c >= regions.size()) continue;
     GRegion *gr = regions[c];
     MVertex *vv[4];
